Move metrics tracker into plugin reporting
